chore: refactored handleCreateExprValues, handleCreateOperators and handleCreateVariableNames into handle-create-construct-funcs.ts, and split out handle-equation-compose.ts into hassign-law-logic dir
